objects: refactor to separate files, add camelize
