+++ src/utils/objects/index.ts
@@ -1,0 +1,13 @@
+import { path } from './path'
+import { propertyOf } from './propertyOf'
+import { isEmpty } from './isEmpty'
+import { setInPath } from './setInPath'
+import { camelize } from './camelize'
+
+export const Objects = {
+  isEmpty,
+  path,
+  propertyOf,
+  setInPath,
+  camelize,
+}
