+++ src/utils/objects/propertyOf.ts
@@ -1,0 +1,1 @@
+export const propertyOf = <T>(name: keyof T): keyof T => name
